+++ src/test/java/org/cicirello/search/evo/BoltzmannSelectionTests.java
@@ -28,6 +28,8 @@ import static org.junit.jupiter.api.Assertions.*;
  */
 public class BoltzmannSelectionTests {
 	
+	// Tests for the bias functions
+	
 	@Test
 	public void testConstantBoltzmannBiasFunction() {
 		ConstantBoltzmannBiasFunction bias = new ConstantBoltzmannBiasFunction(1.0);
@@ -242,6 +244,8 @@ public class BoltzmannSelectionTests {
 			assertEquals(Math.exp(mult*f), bias.bias(f));
 		}
 	}
+	
+	// TESTS FOR THE STANDARD FITNESS PROPORTIONAL VERSION
 	
 	@Test
 	public void testBoltzmannSelectionExponential() {
@@ -1008,6 +1012,776 @@ public class BoltzmannSelectionTests {
 		IllegalArgumentException thrown = assertThrows( 
 			IllegalArgumentException.class,
 			() -> new BoltzmannSelection(0.0)
+		);
+	}
+	
+	// TESTS FOR THE SUS VERSION
+	
+	@Test
+	public void testBoltzmannStochasticUniversalSamplingExponential() {
+		double[] fitnesses = {0, 1, 2, 3, 4, 5};
+		PopulationFitnessVector.Double vector = PopulationFitnessVector.Double.of(fitnesses.clone());
+		
+		BoltzmannStochasticUniversalSampling selection = new BoltzmannStochasticUniversalSampling(1.0, 0.1, 0.5, false);
+		double[] weightedSum = selection.computeWeightRunningSum(vector);
+		double expected = 0.0;
+		double div = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		BoltzmannStochasticUniversalSampling split = selection.split();
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		// Use split copy
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		// reinitialize and use original
+		selection.init(500);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		IllegalArgumentException thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(10.0, 0.0, 1.0, false)
+		);
+		thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(0.09, 0.1, 1.0, false)
+		);
+		thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(5.0, 0.1, 0.0, false)
+		);
+		thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(5.0, 0.1, 1.0, false)
+		);
+	}
+	
+	@Test
+	public void testBoltzmannStochasticUniversalSamplingExponentialInteger() {
+		int[] fitnesses = {0, 1, 2, 3, 4, 5};
+		PopulationFitnessVector.Integer vector = PopulationFitnessVector.Integer.of(fitnesses.clone());
+		
+		BoltzmannStochasticUniversalSampling selection = new BoltzmannStochasticUniversalSampling(1.0, 0.1, 0.5, false);
+		double[] weightedSum = selection.computeWeightRunningSum(vector);
+		double expected = 0.0;
+		double div = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		BoltzmannStochasticUniversalSampling split = selection.split();
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		// Use split copy
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		// reinitialize and use original
+		selection.init(500);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div *= 0.5;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+	}
+	
+	@Test
+	public void testBoltzmannStochasticUniversalSamplingLinear() {
+		double[] fitnesses = {0, 1, 2, 3, 4, 5};
+		PopulationFitnessVector.Double vector = PopulationFitnessVector.Double.of(fitnesses.clone());
+		
+		BoltzmannStochasticUniversalSampling selection = new BoltzmannStochasticUniversalSampling(4.0, 0.1, 1.0, true);
+		double[] weightedSum = selection.computeWeightRunningSum(vector);
+		double expected = 0.0;
+		double div = 4;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		BoltzmannStochasticUniversalSampling split = selection.split();
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		// Use split copy
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 4;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		// reinitialize and use original
+		selection.init(500);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 4;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		IllegalArgumentException thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(10.0, 0.0, 1.0, true)
+		);
+		thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(0.09, 0.1, 1.0, true)
+		);
+		thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(5.0, 0.1, 0.0, true)
+		);
+	}
+	
+	@Test
+	public void testBoltzmannStochasticUniversalSamplingLinearInteger() {
+		int[] fitnesses = {0, 1, 2, 3, 4, 5};
+		PopulationFitnessVector.Integer vector = PopulationFitnessVector.Integer.of(fitnesses.clone());
+		
+		BoltzmannStochasticUniversalSampling selection = new BoltzmannStochasticUniversalSampling(4.0, 0.1, 1.0, true);
+		double[] weightedSum = selection.computeWeightRunningSum(vector);
+		double expected = 0.0;
+		double div = 4;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		BoltzmannStochasticUniversalSampling split = selection.split();
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		// Use split copy
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 4;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		
+		// reinitialize and use original
+		selection.init(500);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 4;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div -= 1.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		div = 0.1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 0.0;
+		for (int i = 0; i < fitnesses.length; i++) {
+			expected += Math.exp(fitnesses[i]/div);
+			assertEquals(expected, weightedSum[i]);
+		}
+	}
+	
+	@Test
+	public void testBoltzmannStochasticUniversalSamplingConstant() {
+		double[] fitnesses = {0, 1, 2, 3, 4, 5};
+		PopulationFitnessVector.Double vector = PopulationFitnessVector.Double.of(fitnesses.clone());
+		
+		BoltzmannStochasticUniversalSampling selection = new BoltzmannStochasticUniversalSampling(1.0);
+		double[] weightedSum = selection.computeWeightRunningSum(vector);
+		double expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		BoltzmannStochasticUniversalSampling split = selection.split();
+		selection.init(1000);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		split.init(1000);
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		
+		selection = new BoltzmannStochasticUniversalSampling(0.5);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(2*fitnesses[i+1]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(2*fitnesses[i+1]);
+		}
+		selection.init(1000);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(2*fitnesses[i+1]);
+		}
+		
+		selection = new BoltzmannStochasticUniversalSampling(2.0);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(0.5*fitnesses[i+1]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(0.5*fitnesses[i+1]);
+		}
+		selection.init(1000);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(0.5*fitnesses[i+1]);
+		}
+		
+		IllegalArgumentException thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(0.0)
+		);
+	}
+	
+	@Test
+	public void testBoltzmannStochasticUniversalSamplingConstantInteger() {
+		int[] fitnesses = {0, 1, 2, 3, 4, 5};
+		PopulationFitnessVector.Integer vector = PopulationFitnessVector.Integer.of(fitnesses.clone());
+		
+		BoltzmannStochasticUniversalSampling selection = new BoltzmannStochasticUniversalSampling(1.0);
+		double[] weightedSum = selection.computeWeightRunningSum(vector);
+		double expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		BoltzmannStochasticUniversalSampling split = selection.split();
+		selection.init(1000);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		split.init(1000);
+		weightedSum = split.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(fitnesses[i+1]);
+		}
+		
+		selection = new BoltzmannStochasticUniversalSampling(0.5);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(2*fitnesses[i+1]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(2*fitnesses[i+1]);
+		}
+		selection.init(1000);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(2*fitnesses[i+1]);
+		}
+		
+		selection = new BoltzmannStochasticUniversalSampling(2.0);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(0.5*fitnesses[i+1]);
+		}
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(0.5*fitnesses[i+1]);
+		}
+		selection.init(1000);
+		weightedSum = selection.computeWeightRunningSum(vector);
+		expected = 1;
+		for (int i = 0; i < fitnesses.length; i++) {
+			assertEquals(expected, weightedSum[i]);
+			if (i < fitnesses.length - 1) expected += Math.exp(0.5*fitnesses[i+1]);
+		}
+		
+		IllegalArgumentException thrown = assertThrows( 
+			IllegalArgumentException.class,
+			() -> new BoltzmannStochasticUniversalSampling(0.0)
 		);
 	}
 }